Misc. bugfixes including memory errors in typemaps (#8)
* Misc. bugfixes including memory errors in typemaps, repaired and new unit tests

* Added SWIG_PYTHON_CAST_MODE for compiling to fix a quirk where function calls fail with TypeError if you use a np.int32 where an int (or np.int64) is expected.

* Improved alias management; expanded compile limits for reading comments from kernels

* Auto-generated files added to .gitignore; unit test correction where time system setting was left altered; better handling of aliases, especially when the given NAIF ID is int32; other misc. bug fixes

* requested files deleted

* Fix typo

* Add instructions for generating a release

Simplify the release process

* Ready for review

* Apply suggestions from code review


* Updated for Rob's changes

* Change Linux builder to newer version

* Forgot cspyce0_info.py

* Completely redo public-to-pypi.

Linux is too different from the others.
Found that multiple pythons can be in the same image.
