+++ .gitignore
@@ -8,3 +8,8 @@ dist/
 cspyce.egg-info/
 **/*wrap.c
 wheelhouse/
+.DS_Store
+cspyce/cspyce0.py
+cspyce/cspyce0_info.py
+swig/vectorize.i
+prerelease_version.txt
