+++ setup.py
@@ -90,7 +90,8 @@ def get_c_libraries():
 
 
 def get_extensions():
-    cspyce_cflags = ['/nowarn'] if IS_WINDOWS else ['-Wno-incompatible-pointer-types']
+    cspyce_cflags = ['/nowarn', '/DSWIG_PYTHON_CAST_MODE'] if IS_WINDOWS \
+        else ['-Wno-incompatible-pointer-types', '-DSWIG_PYTHON_CAST_MODE']
     include_dirs = [os.path.join(cspice_directory, "include"), numpy.get_include()]
 
     cspyce0_module = Extension(
@@ -114,9 +115,18 @@ class MyBuildExt(build_ext):
 
 
 def do_setup():
+    try:
+        directory = os.path.dirname(os.path.abspath(__file__))
+        prerelease_version_file = os.path.join(directory, "prerelease_version.txt")
+        with open(prerelease_version_file, "r")  as f:
+            prerelease_version = f.read().strip()
+            if prerelease_version == 'release':
+                prerelease_version = ''
+    except IOError:
+        prerelease_version = ''
     setup(
         name='cspyce',
-        version='2.0.4',
+        version='2.0.5' + prerelease_version,
         author="Mark Showalter/PDS Ring-Moon Systems Node",
         description="Low-level SWIG interface to the CSPICE library",
         ext_modules=get_extensions(),
