+++ unittests/unittester.py
@@ -2,9 +2,13 @@
 # unittester.py: global unit-tester
 ################################################################################
 
-from unittester_errors     import *
-from unittester_kernels    import *
-from unittester_nokernels  import *
+from test_errors    import *
+from test_kernels   import *
+from test_nokernels import *
+
+from test_cyl_lat_sph import *
+from test_k           import *
+from test_l           import *
 
 ################################################################################
 # To run all unittests...
